Path: Upgrade `Extensions` feature
- Make current `Extensions` feature independent.
- Add `All Access Extension` feature to current `Extensions` feature.
- Apply improved, upgraded `Extensions` feature to the Adaptive operation.
- Apply `Use outline` feature found in PocketShape to Adaptive operation.
