+++ src/Mod/Path/PathScripts/PathAdaptive.py
@@ -39,6 +39,13 @@ import math
 import area
 from pivy import coin
 
+# lazily loaded modules
+from lazy_loader.lazy_loader import LazyLoader
+Part = LazyLoader('Part', globals(), 'Part')
+FeatureExtensions = LazyLoader('PathScripts.features.PathFeatureExtensions',
+                                globals(),
+                                'PathScripts.features.PathFeatureExtensions')
+
 __doc__ = "Class and implementation of the Adaptive path operation."
 
 def convertTo2d(pathArray):
@@ -340,12 +347,7 @@ def Execute(op,obj):
         if obj.Tolerance < 0.001:
             obj.Tolerance = 0.001
 
-        pathArray = []
-        for base, subs in obj.Base:
-            for sub in subs:
-                shape = base.Shape.getElement(sub)
-                for edge in shape.Edges:
-                    pathArray.append([discretize(edge)])
+        pathArray = _get_working_edges(op, obj)
 
         #pathArray=connectEdges(edges)
         path2d = convertTo2d(pathArray)
@@ -469,13 +471,78 @@ def Execute(op,obj):
         job.ViewObject.Visibility = oldJobVisibility
         sceneClean()
 
+def _get_working_edges(op, obj):
+    pathArray = list()
+    edge_data = list()
+    edge_list = list()
+
+    for base, subs in obj.Base:
+        for sub in subs:
+            if obj.UseOutline:
+                face = base.Shape.getElement(sub)
+                shape = Part.Face(face.Wires[0])
+            else:
+                shape = base.Shape.getElement(sub)
+
+            # Apply All Access Extension feature when enabled
+            aa_value = obj.AllAccessExtension.Value
+            if aa_value > 0.0:
+                aa_face = FeatureExtensions._get_all_access_face(obj, 
+                                                            base.Shape,
+                                                            shape,
+                                                            aa_value,
+                                                            discretize_factor=0.5)
+                if aa_face:
+                    shape = aa_face
+            # Eif All Access Extension
+
+            # Only add edges not found in extensions wires earlier
+            for e in shape.Edges:
+                mp = _get_edge_midpoint(e)
+                edge_data.append(mp)
+                edge_list.append(e)
+    # Efor
+
+    # add edges from active extensions
+    op.exts = [] # pylint: disable=attribute-defined-outside-init
+    for ext in FeatureExtensions.getExtensions(obj):
+        wire = ext.getWire()
+        if wire:
+            face = Part.Face(wire)
+            op.exts.append(face)
+            for e in face.Edges:
+                mp = _get_edge_midpoint(e)
+                if mp in edge_data:
+                    # Edge exists. Remove it.
+                    i = edge_data.index(mp)
+                    edge_data.pop(i)
+                    edge_list.pop(i)
+                else:
+                    edge_data.append(mp)
+                    edge_list.append(e)
+
+    for edge in edge_list:
+        pathArray.append([discretize(edge)])
+
+    return pathArray
+
+def _get_edge_midpoint(e):
+    precision = 4
+    midpnt = e.valueAt(e.FirstParameter + (e.Length / 2.0))
+    mpx = round(midpnt.x, precision)
+    mpy = round(midpnt.y, precision)
+    mp = 'x{}_y{}'.format(mpx, mpy)
+    return mp
+   
 
 class PathAdaptive(PathOp.ObjectOp):
     def opFeatures(self, obj):
         '''opFeatures(obj) ... returns the OR'ed list of features used and supported by the operation.
         The default implementation returns "FeatureTool | FeatureDepths | FeatureHeights | FeatureStartPoint"
         Should be overwritten by subclasses.'''
-        return PathOp.FeatureTool | PathOp.FeatureBaseEdges | PathOp.FeatureDepths | PathOp.FeatureFinishDepth | PathOp.FeatureStepDown | PathOp.FeatureHeights | PathOp.FeatureBaseGeometry | PathOp.FeatureCoolant
+        return PathOp.FeatureTool | PathOp.FeatureBaseEdges | PathOp.FeatureDepths \
+            | PathOp.FeatureFinishDepth | PathOp.FeatureStepDown | PathOp.FeatureHeights \
+            | PathOp.FeatureBaseGeometry | PathOp.FeatureCoolant | PathOp.FeatureLocations
 
     def initOperation(self, obj):
         '''initOperation(obj) ... implement to create additional properties.
@@ -491,7 +558,6 @@ class PathAdaptive(PathOp.ObjectOp):
         obj.addProperty("App::PropertyDistance", "LiftDistance", "Adaptive", "Lift distance for rapid moves")
         obj.addProperty("App::PropertyDistance", "KeepToolDownRatio", "Adaptive", "Max length of keep tool down path compared to direct distance between points")
         obj.addProperty("App::PropertyDistance", "StockToLeave", "Adaptive", "How much stock to leave (i.e. for finishing operation)")
-        # obj.addProperty("App::PropertyBool", "ProcessHoles", "Adaptive","Process holes as well as the face outline")
 
         obj.addProperty("App::PropertyBool", "ForceInsideOut", "Adaptive","Force plunging into material inside and clearing towards the edges")
         obj.addProperty("App::PropertyBool", "FinishingProfile", "Adaptive","To take a finishing profile path at the end")
@@ -514,6 +580,13 @@ class PathAdaptive(PathOp.ObjectOp):
         obj.addProperty("App::PropertyAngle", "HelixAngle", "Adaptive",  "Helix ramp entry angle (degrees)")
         obj.addProperty("App::PropertyLength", "HelixDiameterLimit", "Adaptive", "Limit helix entry diameter, if limit larger than tool diameter or 0, tool diameter is used")
 
+        if not hasattr(obj, "UseOutline"):
+            obj.addProperty("App::PropertyBool",
+                            "UseOutline",
+                            "Adaptive",
+                            "Uses the outline of the base geometry.")
+
+        FeatureExtensions.initialize_properties(obj)
 
     def opSetDefaultValues(self, obj, job):
         obj.Side="Inside"
@@ -521,7 +594,6 @@ class PathAdaptive(PathOp.ObjectOp):
         obj.Tolerance = 0.1
         obj.StepOver = 20
         obj.LiftDistance=0
-        # obj.ProcessHoles = True
         obj.ForceInsideOut = False
         obj.FinishingProfile = True
         obj.Stopped = False
@@ -533,13 +605,14 @@ class PathAdaptive(PathOp.ObjectOp):
         obj.StockToLeave = 0
         obj.KeepToolDownRatio = 3.0
         obj.UseHelixArcs = False
+        obj.UseOutline = False
+        FeatureExtensions.set_default_property_values(obj, job)
 
     def opExecute(self, obj):
         '''opExecute(obj) ... called whenever the receiver needs to be recalculated.
         See documentation of execute() for a list of base functionality provided.
         Should be overwritten by subclasses.'''
         Execute(self,obj)
-
 
 
 def Create(name, obj = None):
